Improve duplicate embed prevention
Set `window.hypothesisHasRun` rather than trying to detect duplicate
embeddings in the background script. For the extension this should be
very reliable because the symbol will be isolated to the content script
context and can never conflict with variables set by page authors.

Don't evaluate the yepnope script except when it is actually needed.