Isolate static assets in h:static
There are a number of reasons to isolate the static assets into their
own directory.

- An edge case in Pyramid: https://github.com/Pylons/pyramid/pull/1377
- Make serving them with a reverse proxy safer by ensuring that no code
  or configuration lives beneath the static directory
- Cleanliness

To avoid a big move happening shortly after this, also rename some of
the asset subdirectories to close #1317.

- js -> scripts
- css -> stylesheets
- lib -> scripts/vendor
- lib/images -> images
- browser/chrome/js/background.js -> browser/chrome/background.js
- browser/chrome/public/js/destroy.js -> browser/chrome/public/destroy.js

Finally, the locale directory was totally busted. The annotator.po file
included is the defualt locale, and it should have been .pot. Since it
includes only the default strings, there's no reason to serve it. I've
stripped it and we'll reinstate it correctly later.

Close #1311