[chrome] run embed code in page context
Rather than running the embed code in the content script context and
duplicating the double embed prevention in the extension, run the
embed code in the page context since it now handles the double embed
prevention itself.

This is also simpler to reason about because the embed code runs in
a similar way for all injection methods (boomarklet, script embed, or
extension).

Close #2194