Remove unnecessary polyfills from production
Function.prototype.bind is supported in every browser for a long time.
It's only needed because, oddly, PhantomJS 1.8 doesn't support it.

The window.URL constructor is supported in IE10+ and every other major
browser for a while now. It's also not supported in PhantomJS 1.8,
though. The application currently doesn't work on IE < 10 anyway so
let's just remove this for now.