Show num. annotations on browser badge
Each time the Chrome user loads a web page, send a request to our API
for the number of annotations of that page. If there are > 0 annotations
then show the number of annotations as a badge on the Chrome extension
button.

This requires the Chrome extension code to have access to the URL of our
API, which it doesn't currently. So have buildext.py write the URL into
a settings.json file that's shipped with the extension, and add
settings.js / h.Settings to the extension code to wrap this JSON file.

Also move the existing blocklist setting into settings.js/json.