Add a Chrome badge blocklist
Add a blocklist of URLs that the Chrome extension's badge will never
show a number for.

This is done by adding a separate API for the Chrome extension to use to
get its badge number, and this API just returns 0 for blocklisted URLs.
Doing it this way means that the blocklist is server-side, so it can be
stored in the database and editable by an admin page.

The new API view for the blocklist accesses the badge model from the
badge package which is part of the app not part of the API, so the view
is also be part of the app not part of the API.

This means that it has to import a bunch of names from the API though.