Fix and simplify restoration of tab state after extension reload
The TabStore documentation claimed that it was used to restore
H state for tabs across reload. This is false because the storage
used tab IDs as keys and they do not persist across browser sessions.

For persisting tab state across extension reloads, reload tab state
from storage in the install event handler and invoke
onTabStateChanged() to refresh the tabs from the current state
in TabState.