Test for the 'ready' flag on TabState explicitly when deciding whether to inject the sidebar
The 'ready' state flag was added to make the state for
a tab different when the tab was active but not yet loaded
vs. active and loaded so that `onTabStateChanged` would be triggered.

However, the 'ready' flag was not actually read anywhere
which was confusing. This commit replaces the call to Chrome APIs
to get the tab loaded state in `updateTabDocument` with a check
of the `ready` flag instead.

This is also a small step towards earlier loading of the
sidebar in future by not waiting for the complete page load
(including all 1st and 3rd-party resources) before injecting it.