Fix a bug where the tabs that errored were not always re-activated after navigation
When navigating from one page to another within a tab,
H tries to maintain the active/inactive state for a tab.

If a user tried to activate H but it failed and they then
navigate to another page, H would try to re-activate
the extension. This did not happen in the case where Chrome
_replaced_ the tab with a pre-rendered one rather than
updating the existing one however.