Remove the incomplete Firefox extension
A recent successful prototype of a Firefox extension
based on the new WebExtensions API highlighted that as
the way forwards for Hypothesis on Firefox, as it lets
us offer feature parity with much easier maintenance.

If we decide that we want to support the extension for Firefox pre-45, we may
do that by shimming the Chrome APIs using the JPM SDK APIs, similar
to what https://github.com/einaregilsson/Redirector have done.

This commit removes the legacy incomplete
Firefox extension which is missing all of the recent improvements
and tests for the Chrome extension.