Fix some Chrome extension tests in Chrome
The code has been refactored to use fetch() instead of XMLHttpRequest,
but the tests still use sinon.FakeServer which patches XMLHttpRequest.

The tests still pass in PhantomJS (and so on TravisCI) because PhantomJS
has a polyfill for fetch() that calls XMLHttpRequest. But Chrome's
fetch() does not call XMLHttpRequest so the tests fail in Chrome.

Fix the tests to stub fetch() instead of using FakeServer.