Send cookies in badge API requests
A recent refactoring of the Chrome extension code that makes the HTTP
requests to the badge API to use fetch() instead of XMLHttpRequest
inadvertently changed the request to _not_ include cookies, meaning that
private and group annotations were no longer counted by the badge.

Put the cookies back into the request.

Fixes #2928.