Correctly generate bouncer match URL in manifest
Match URLs in the Chrome extension manifest file must contain wildcard
characters. We were incorrectly generating a URL that referenced only
the top-level of `https://hpt.is/`.

In addition, it seems safer to always generate the
`externally_connectable` clause in the manifest, as this means that
external connectability for *extensions* (not web pages) is the same
regardless of whether a bouncer URL is provided. See:

  https://developer.chrome.com/extensions/manifest/externally_connectable#without-externally-connectable