Use only one mechanism for indicating first-run state in the /welcome page
Depending on which route the Help page was loaded from, it used two
different mechanisms to read the config info.

This was presumably done so that the config info would be visible
to annotator/main.js if loaded in a background page.

We're not actively doing that however, so this removes that
option for the time being.