Preserve URL fragments when redirecting to PDF.js in extensions
Direct linking to annotations in PDFs was broken on systems with the
Chrome extension installed because the '#annotations' fragment was lost
when the Chrome extension redirected from the original URL to the
bundled PDF.js viewer.

This is fixed by extracting and URI-encoding the non-fragment parts of
the URL when generating the 'file' query string param for the PDF.js
viewer and then appending the original fragment to the resulting URL,
which becomes:

  chrome-extension://<ID>/content/viewer.html?file=<...>#annotations:ID

When the embedded version of H in the bundled PDF.js viewer loads, it
will then pick up the fragment and scroll to and focus on the
annotation.