Make direct-link injection robust to URL fragment changes as the page loads.
Sites such as Medium and EBSCOHost modify the URL fragment as the page
loads, causing direct-link annotation ID fragments to be lost by the
time the app's init code runs.

When a tab's URL changes, extract the '#annotations' fragment and pass
it to the application as configuration info when the app is injected.
If the URL fragment is removed or replaced by client-side JS as the page
loads, the app will still get the original annotation ID that was
captured by the extension prior to the page load completing.

The direct-linked ID is remembered until the sidebar is successfully
injected, at which point it is cleared so that it is not re-used on the
next navigation within the tab.