Add options page to extension. Allow disabling badge.
Add an options pane to the Chrome extension, with the first available
option being the ability to disable the annotation count badge.

The settings are persisted in a dictionary using the `chrome.storage`
API, and used to disable the code responsible for updating the badge
state when a tab changes.

This adds the "storage" permission, which does not result in additional
install-time warnings [1].

[1]: https://developer.chrome.com/extensions/permission_warnings#nowarning