Add a straightforward build for the Chrome extension
Rather than moving `buildext.py` as-is across to this repository, with
all of its attendant Pyramid baggage, it made sense to me to try to
recreate the Chrome extension build rather more simply.

This commit adds most of (but by no means all) of the build steps
necessary to bundle the Chrome extension in a single `make` invocation.
Settings for the extension come from a configurable SETTINGS_FILE, which
defaults to the `settings.json` at the root of the repository.