Build extension bundle with browserify
Add tasks to the Makefile so that the extension is built into a bundle
by browserify.

In order to make it clear what is a "build output" vs a distribution
package, I've opted to put build outputs in the `build/` directory and
distribution packages in `dist/`.

Again in order to clarify that files in different locations come from
different places, I've moved the target of `extension.bundle.js` to the
extension root rather than inside the `lib/` directory.