Make the Makefile less Chrome-specific
The way we're building extensions now lends itself directly to building
Chrome and Firefox extensions using the same toolchain, altering only
the settings file in use.

As such, it makes sense for the Makefile to not be quite as
Chrome-specific as it was previously. This commit simplifies a few
things and makes it possible to build Firefox extensions.