Add script to update PDF.js to latest build from PDF.js viewer page
As described at https://github.com/mozilla/pdf.js/wiki/Setup-pdf.js-in-a-website#from-examples
the latest build of the viewer is always available from PDF.js' GitHub Pages
site.

Add a script which fetches the latest build and copies the relevant
parts to src/vendor/pdfjs.

The PDF.js build is moved from src/chrome/content to src/vendor/pdfjs to
make it more obvious that this is vendor code.