Use service URL from settings to construct welcome page URL
Use the service URL from the extension's settings to construct the URL
for the '/welcome' page and the URL pattern that allows requests made by
the extension to set cookies for the Hypothesis domain.

Previously both of these were hardcoded to 'https://hypothes.is', an
issue that was noticed when testing deployment of the service to
Skyliner under a different domain.