Add a note about the misleading error message in devtools
Add a note about an issue that confused me while updating the extension
to use the client's boot script. The Network tab in Chrome devtools may
report that the sidebar app HTML file failed to load, even though it
actually loads successfully. If sidebar app is not visible, this may be
because of a problem when app.html loads rather than a problem with the
request for the HTML file.