Configure PDF viewer to load client assets from extension
The sidebar injector code currently only injects config script tags when
injecting the client into HTML documents, therefore the code to load the
client's assets from the extension added in previous commits did not
apply. Passing the same config to the PDF viewer will require additional
work because injecting the client into a PDF involves first
loading a new page into the tab (the PDF.js viewer), waiting for that to
load and _then_ injecting the config and boot script.

A simple solution for the moment is just to hardcode the asset URL and
sidebar app URL to '/client/' and '/client/app.html' respectively.