Don't do git checkout or pull from a script
The docs for releasing a Chrome extension say to checkout and pull the
latest master branch manually.

The script as its currently written assumes that the user's master
branch is setup as a tracking branch of the hypothesis organization's
master branch on GitHub, which while a common setup isn't always the
case (it's not the case for me, for example).

For example it's common to fork a github repo and then clone your fork.

Replace the git checkout master and git pull with some code to crash if
the local git repo's HEAD is not the same commit as the HEAD of the
github.com/hypothesis/browser-extension repo.