Make return value of boot script serializable
The boot script is a Browserify bundle and the result of the last
(and only) top-level statement in the script is not serializable.

As a result, when the browser extension tries to execute the boot script
as a content script in Firefox, it reports an error when it tries to
return the result of the script to the background page. Chrome on the
other hand just returns `null` in this case.

This commit works around the problem by appending a serializable
expression (just `null`) to the end of the bundle. If in future other
scripts from the client are executed as content scripts then they too
will need this workaround.