Enable ES6 usage
Enable use of ES6 language features using Babelify.

Since the code in this repository only needs to target reasonably recent
versions of Chrome and Firefox, use babel-preset-env to control which
ES6 => ES5 transformations are applied based on the browser versions we
want to support.

The versions are initially set conservatively to the minimum Firefox
version that has sufficiently complete WebExtensions support (48) and
Chrome circa ~2 years ago. We can bump these in future when we get data
that gives us a better idea of what versions our users are actually
using.