Add OAuth configuration to extension settings
OAuth is currently disabled but enabling it once everything else is
ready will be a matter of flipping the "oauthEnabled" flag.

Note that this does not yet work for the stage and prod extensions when
built locally because the extension IDs are currently different than
when using an extension from the Chrome Web Store.