Allow extension settings to specify the extension ID
In order for the service's OAuth configuration to work with a locally
built production or staging extension, the extension ID must match that
of the pre-registered OAuth clients for the Chrome extension.

The extension ID is usually different on each developer's system, but
can be fixed by specifying the "key" property in the extension's
manifest. See https://developer.chrome.com/apps/manifest/key