Use fixed ID for locally build Chrome prod and stage extensions
Add "key" manifest settings from the prod and stage extensions from the
Chrome web store to the local build settings.

This ensures that locally built Chrome prod & stage extensions have the
same IDs as the versions from the Chrome Web Store.

This comes with the disadvantage that it is no longer possible to have
the "real" Hypothesis Chrome extension installed alongside a
"development" one.