Dissable badge by default
Dissable the badge by default and on firefox dissable it permanently
(as there is no setting to enable it). In order to do this we need
to add a new badgeEnable state param to the store, otherwise any
previously installed extension settings will override the default
settings of the latest version and the badge won't be dissabled.
The badge is being changed to an opt in setting because it accounts
for 90% of our transactions and is a scallabilty problem. Every
page the user browses to when the badge is enabled causes a
transaction to es and our db.