Fix `make` exit status if browserify build fails
If the browserify build step of `make build/extension.bundle.js` failed,
the make sequence continued and exited with a zero status. This caused
the Travis build to pass even though it should have failed.

This happens because the `browserify ... | exorcist ...` pipeline exits
with the status of the last command in the pipeline, and exorcist exits
with a zero status if Browserify fails and the input is empty.

Splitting up the pipeline into two steps fixes the issue as make stops
when Browserify exits with a non-zero status.