Support direct links to groups
Support the `#annotations:group:<ID>` syntax used to direct-link to
groups.

As part of this, refactor the code so that the knowledge about the
different kinds of direct-link query that can be used only exists in
one place, `direct-link-query.js`, which is responsible for mapping the
query to additional client configuration.

Also add some comments about why the browser extension has to deal with
direct-links itself in the first place.