Convert proxyquire in install-test to mockable-imports
A slight change to src/common/install.js was required because
mockable-imports does not support installing mocks until after the
module is evaluated.

In Chrome, install.js should initialize automatically. In the tests
however it needs to wait until we call it. Therefore wrap the init code
in an exported function, which is called automatically outside of tests.