Fix detection of whether a tab is using Chrome's PDF viewer
Chrome recently made changes to how the native PDF viewer is loaded [1],
which slightly changes the HTML of the top-level document in tabs that
are using this viewer. Specifically, the `<embed>` element no longer has a
`name` attribute. This broke the browser extension's detection of
whether the current document is a PDF by checking whether Chrome is
using its native PDF viewer to show it [2].

Resolve the issue by adjusting the CSS selector slightly.

To test, verify that activating the extension on a PDF uses the bundled
PDF.js viewer both when the "MimeHandlerViewer in cross-process frame"
feature flag in Chrome is turned on (at chrome://flags) and when it is
turned off.

Fixes #212

[1] https://bugs.chromium.org/p/chromium/issues/detail?id=659750

[2] nb. We use this method rather than relying only on the file
    extension because not all PDF URLs have a path with a `.pdf` suffix,
    and AFAIK there is no way to get the MIME type of an already-loaded
    document.