Force the python env version to 2.7
The `web-ext` npm dependency installs `dtrace-provider` via the `web-ext >
bunyan > dtrace-provider` transitive dependency. dtrace-provider
requries node-gyp to build and, except for the very latest version of
node-gyp, this fails to build with Python 3 and produces a ton of
confusing debug output when running `npm install` if the `python` binary
is Python 3.

dtrace-provider uses a stub implementation if it fails to build, so the
issue isn't fatal but it is very confusing. We don't even need this
package, but npm doesn't provide a convenient way to override it AFAIK.

Since we use pyenv for all of Hypothesis' projects to manage the Python
version, use a .python-version file to force the use of Python 2.7 here
as a workaround.

See https://github.com/hypothesis/browser-extension/pull/246 for
the full saga.