Trace deployment steps
When a deployment step fails, this should make it easier to identify
which step was affected.

Note that since the deploy script references secrets, this relies on the
CI system to suppress these secrets in build logs, which both Travis and
GitHub Actions do.

This should make it easier to identify which step o