Prepare test modules for conversion to ES modules
When these modules are converted to ES modules, `$imports` will become a
separate export rather than a property of the default import.

Also fix an issue where the mockable-imports plugin was processing files
in the `tests/` directory, which is a) unnecessary and b) causes an
infinite recursion error for reasons I have yet to debug.