Add Jenkins build parameter to update Hypothesis client
Add a `BUILD_TYPE` parameter that controls whether Jenkins does a
regular build or updates the Hypothesis client and pushes a new tag.

When building with the "update-hypothesis-client" build type, the Test +
Build steps are replaced with a step that installs the latest version of
the "hypothesis" npm dependency and then creates a new npm version, as
well as creating a new tag. The changes are pushed back to the source
branch, which will always be "master", which will in turn kick off
another Jenkins build which will actually test and deploy the new
Hypothesis client release.

Remove the `tools/update-client` script has this is a manual tool for
updating the Hypothesis client which is superceded by the Jenkins task
to do this.