Reduce the number of badge requests
It implements a custom debouncing mechanism.

A more standard approach like using `lodash._debounce` is not suited for this
particular case, because TabState is shared amoung all the tabs.

Reloading two tabs simulatneously shouldn't make one of the requests to be
cancelled.

The cancellation of a pending request could happen during the waiting period or
in-flight badge request.

Resolves #396