Reduce number of badge requests by fetching when tab is complete
The `tab.onUpdated` event is fired many times during the loading
of a page. The larger the document the larger the number of times
the event is triggered.

By moving the method that request badges to a place that is only
reached when the page is fully loaded, we avoid unnecessary calls.