Fix error when deactivating extension a second time in a tab
Fix an error about a duplicate `annotatorLink` variable when
de-activating the extension in a tab where it had previously been
active.

It appears that injected scripts all share the same JS global, so if the
`unload-client.js` script is injected multiple times then subsequent
runs would fail due to the `annotatorLink` variable having already been
declared. This didn't happen in the past because the `const` was
transpiled down to `var`, which allows redeclaration. This no longer
happens.