Fix PDF.js viewer failing to load without unsafe-eval
Fix loading of the PDF.js viewer (viewer.html) after removing the
unsafe-eval permission from the extension's content security policy.

It turns out that an async/await polyfill (regenerator) in PDF.js's
source bundles relied on this in a fallback code path if assigning a
global `regeneratorRuntime` variable failed.

Fix the issue by predefining the `regeneratorRuntime` global before
PDF.js loads. We should be able to remove this in future after upgrading
to a release of PDF.js that uses native async/await.

 - Add `pdfjs-setup-env.js` as a place to prepare the global environment
   before PDF.js loads and predefine `regeneratorRuntime` in it

 - Load pdfjs-setup-env.js before PDF.js in viewer.html

 - Configure PDF.js to use a custom wrapper around pdf.worker.js in
   pdfjs-init.js. This wrapper, in pdfjs-worker-init.js, loads
   pdfjs-setup-envjs before the real worker script.