Support injecting client into VitalSource Bookshelf
Unlike other HTML pages, the client needs to be injected into a specific
non-main frame of this application.

The current implementation requires an extra `chrome.webNavigation`
permission to enumerate frames in the active tab. Requesting this
permission displays a scary, and in this context inaccurate, prompt. We
therefore only request it when activating the extension on a VitalSource
tab.