Modernize and add types to `promisify` utility
Modernize and add types [1] to the utility that we use to convert
callback-returning Chrome extension APIs to Promise-returning ones.

Adding these types triggered some conflicts in an `injectIntoRemoteDocument`
helper in `sidebar-injector.js` due to different branches of a function
returning incompatible types. Refactoring this function to async/await resolved
this issue by discarding unused information about the results (the caller of
`injectIntoRemoteDocument` only cares about when the operation has completed).

[1] This makes use of variadic tuple types. See
    https://www.typescriptlang.org/docs/handbook/release-notes/typescript-4-0.html