Add types to functions in sidebar-injector.js
This revealed a potential scenario where tabs could be missing ID or URL
information. All normal tabs should have this information available due
to the extension's permissions, but there may be unusual tabs that do
not. The public methods of `SidebarInjector` now check for this and
throw an exception if the fields are missing.