Refactor HelpPage class
The implementation and tests were rather verbose. This commit rewrites HelpPage
to have the same API (at least, the parts that were actually used) and
functionality but a more succinct implementation and tests.

In the process access to the Chrome APIs has been converted to use the
`chrome-api` module and an obsolete code path for old versions of
Firefox was removed.