Support optional `chrome.webNavigation.getAllFrames` use
This API is needed on certain websites (eg. VitalSource Bookshelf) where
we need to inject the client into a specific frame in the tab which is
not the main frame.

Since optional APIs may not exist on the global `chrome` object when
`getChromeAPI` is called, the `webNavigation` property of the returned
object needs to resolve lazily. The caller is responsible for making
sure that the permission has been requested by the time the API is
accessed.