Replace Browserify with Rollup
Replace Browserify with a more modern bundler that understands ES modules
natively and has better ecosystem support.

Apart from the bundler change, there is also a change to how the code is bundled
for tests. Instead of using a Karma framework, we create a bundle ourselves from
the test files and then instruct Karma to load the pre-bundled code. This
eliminates the need for developers to understand a glue dependency
(karma-browserify) and also makes it easier to inspect the generated test
bundle, since it is output to the build directory before the tests run.