Move settings data into the main extension bundle
The settings data for the extension is currently loaded into the background page
as a separate `settings-data.js` script before the main extension bundle. In
Manifest V3 the background code runs in a Service Worker rather than a web page.
The service worker needs to have a single script as the entry point, although it
can be an ES module which can then import other scripts.

To expose the settings to the background page in a way which works in both
Manifest V2 and Manifest V3, incorporate it directly into the bundle by using
Rollup's JSON plugin.

The settings data from `settings/<filename>.json` is now incorporated into the
bundle as follows:

 1. make runs `tools/settings.js` to generate `build/settings.json` based on
    the input specified via `SETTINGS_FILE=<path>`.

 2. `src/background/settings.js` now imports `build/settings.json` directly,
    using Rollup's JSON plugin to translate it to an ES module.