Show original source locations in test error stacktraces
Use the karma-source-map-support package to make assert error stacktraces
reported by Karma show the original source location instead of the
location within the test bundle.

karma-source-map-support is a simple package which loads the
source-map-support package in the browser [1]. This package is
Chrome-specific. In non-Chrome browsers the tests can still run but will
show the location in the test bundle.

[1] https://github.com/evanw/node-source-map-support#browser-support