Add wrappers around extension APIs for executing scripts
One of the major changes between Manifest V2 and Manifest V3 is the API
for executing scripts in a tab. Add `executeFunction` and
`executeScript` wrappers in `chrome-api.js` which wrap the extension
APIs for script execution provided by the current browser. Currently
this calls the `chrome.tabs.executeScript` API from Manifest V2 /
WebExtensions. In future it will call `chrome.scripting.executeScript`
in browsers that support Manifest V3.

As well as abstracting over browser differences, these wrappers also
enable better typing and a simpler interface for other parts of the
extension.