Don't request "webNavigation" permission if we already have it
Calls to inject the client into the VitalSource viewer happen in two
ways:

 1. Manually, in response to a user activating the extension
 2. Automatically, when the browser attempts to re-inject Hypothesis
    after a potential navigation ("potential" because the extension
    doesn't know if the tab URL changed as a result of a client-side URL
    update or a server-side page change).

In scenario (2) the tab would always be flagged as having an error
because the `chrome.permissions.request` API fails if called outside of
a user gesture, even if the extension already has the permission.

Fix the issue by checking for the permission using
`chrome.permissions.getAll`, which does not require a user gesture, and
then only requesting if if needed.