Add `--grep` and `--watch` options to the test command
Similar to the client and LMS projects, I added support for the `--grep`
and `--watch` options for testing.

Because the browser-extension codebase is small it doesn't take much
time to bundle the files, so the `--grep` option is not that super
needed. However, the `--watch` option is a useful one.

In any case, for consistency and developer experience, I think it is
better to have these two options available.

I copy-pasted the `runTest` function from the
`@hypothesis/frontend-build`. I didn't use this package directly because
it has a number of peer dependencies (sass, tailwindcss, etc.) to are
irrelevant to this project.

The test command can be invoke in a number of ways:
* make test ARGS="--watch --grep injector"
* gulp test --watch --grep injector
* yarn gulp test --watch --grep injector