Add `--tags` argument to `git describe`
The git-describe npm package, which is used by tools/settings.js to
generate a version number, defaults to enabling this flag. Without it
only annotated tags are considered, and none of the recently created
tags are annotated. Hence to get the build logs to display a version
number that matches what is in the built packages, we need to add this
flag.