Pass ref from release workflow to CI workflow
When the update-client workflow runs the release workflow, it passes the
new ref created after the Hypothesis client package is updated. The
release workflow was using this ref when running the release scripts,
but not when calling the continuous-integration workflow to actually
build the client. Hence the built client ended up with the wrong version
number. Fix this so that the new ref is passed from update-client to
release and then to the continuous-integration workflow.