Enable messages from bouncer to activate the extension
Add support for a new external message, "activate", which enables the extension
on the tab that sent the message. The existing "ping" message also gains
support for a `queryFeatures` field which allows bouncer to test whether
the installed extension supports the "activate" message.

This enables the bouncer service to activate a specific Hypothesis extension
directly, as opposed to having all installed Hypothesis extensions watch and
respond to the presence of an `#annotations:` fragment in the URL. The latter
approach causes a conflict when multiple Hypothesis extensions are installed.

Part of #1177