Sync client config in PDF.js with config for web pages
The `profileAppUrl` setting was added to `extension.js` but not mirrored in
`pdfjs-init.js`. This caused the wrong profile app URL to be used when viewing a
PDF. It also lead to a cryptic error from the client's `processUrlTemplate`
function when testing with a PDF and a local build of the client, due to
`processUrlTemplate` not being able to handle `chrome-extension://` URLs
properly.