Fix overflow of longer error messages
Error messages of more than about 60 characters would overflow the dark grey
container. Change the `<pre>` to a `<div>` to fix this. This may lead to
word breaks if an error message contains eg. a long URL, but I think that is
acceptable, and preferable in this context to having eg. hidden overflow.