Remove raven-js crash reporting from Service Worker
Remove crash reporting infrastructure from the Service Worker. It was based on
the old raven-js client library for Sentry and was not actually used. To make it
work we would need to migrate to the new Sentry client libraries and do a bunch
of other work.  Since it is not currently functional we can remove it in the
interim and avoid maintaining it.

The Makefile was updated as there are no longer any files matching
src/background/*.js.