Fix tools/update-client for Yarn v3
`yarn add hypothesis@latest` sets the version range for the "hypothesis"
dependency in package.json to "latest", instead of resolving it to a semver.
Consequently the regex to parse the version number failed to match.

`yarn add <dep>` generates a semver range for the `latest` tag by default, so
just use that.