Simplify Chrome API wrapper
Since the migration to Manifest V3 is now complete, all of the `chrome` APIs
support returning promises, so we don't need to wrap the older
callback-returning APIs to return promises. This simplifies the code and
also removes the need for a TypeScript workaround when resolving the type
of wrapped functions.