Disable automatic upload of extension after commits to main
Remove the workflow steps that automatically automatically build and upload the
QA and production extensions after commits to main. The release fails if there
is a draft extension in the pending review state when an extension is uploaded.
This has been happening frequently in recent weeks due to reviews in the Chrome
Web Store having become much slower, for both the QA and production extension.
Previously reviews would often happen in 30 minutes or so, but now sometimes
take several days. Hence automated releases frequently failed.

Instead switch to a process where extension releases are triggered manually:

 1. Go to the Chrome Web Store and verify that the QA and production extensions
    are both released and not in a pending-review state. If either is in a
    pending review state, either wait for the review to be completed or cancel
    the review.

 2. Go to https://github.com/hypothesis/browser-extension/actions/workflows/release.yml
    and run the Release workflow manually.

 3. Wait for the QA extension to be reviewed in the Chrome Web Store and verify
    the extension works as expected.

 4. Publish the production extension in the Chrome Web Store