Correct `authDomain` setting in staging extension
This value needs to match the `authority` field of the `/api/profile` response
in order for first-party users to be treated as first party.  Otherwise some
functionality is disabled, such as the "Log out" option in the account menu.

This has been broken ever since the production and staging databases for h were
split, but somehow we failed to notice.